Remove batch and partition implementations along with their associated tests. Introduce benchmark tests for task execution performance.
